move App::component to App.js
