+++ lib/App.js
@@ -6,11 +6,14 @@
  *
  */
 
+var path = require('path');
 var EventEmitter = require('events').EventEmitter;
 var tracks = require('tracks');
 var util = require('racer/lib/util');
 var derbyTemplates = require('derby-templates');
+var templates = derbyTemplates.templates;
 var documentListeners = require('./documentListeners');
+var components = require('./components');
 var Page = require('./Page');
 var serializedViews = require('./_views');
 
@@ -25,7 +28,7 @@ function App(derby, name, filename, options) {
   this.bundledAt = '{{DERBY_BUNDLED_AT}}';
   this.Page = createAppPage();
   this.proto = this.Page.prototype;
-  this.views = new derbyTemplates.templates.Views();
+  this.views = new templates.Views();
   this.tracksRoutes = tracks.setup(this);
   this.model = null;
   this.page = null;
@@ -172,6 +175,43 @@ App.prototype.serverUse = util.serverUse;
 App.prototype.loadViews = function() {};
 
 App.prototype.loadStyles = function() {};
+
+App.prototype.component = function(viewName, constructor) {
+  if (typeof viewName === 'function') {
+    constructor = viewName;
+    viewName = null;
+  }
+
+  // Inherit from Component
+  components.extendComponent(constructor);
+
+  // Load template view from filename
+  if (constructor.prototype.view) {
+    var viewFilename = constructor.prototype.view;
+    viewName = constructor.prototype.name || path.basename(viewFilename, '.html');
+    this.loadViews(viewFilename, viewName);
+
+  } else if (!viewName) {
+    if (constructor.prototype.name) {
+      viewName = constructor.prototype.name;
+      var view = this.views.register(viewName);
+      view.template = templates.emptyTemplate;
+    } else {
+      throw new Error('No view name specified for component');
+    }
+  }
+
+  // Associate the appropriate view with the component type
+  var view = this.views.find(viewName);
+  if (!view) {
+    var message = this.views.findErrorMessage(viewName);
+    throw new Error(message);
+  }
+  view.componentFactory = components.createFactory(constructor);
+
+  // Make chainable
+  return this;
+};
 
 App.prototype.createPage = function() {
   if (this.page) {
