+++ lib/components.js
@@ -7,18 +7,17 @@
  *
  */
 
-var path = require('path');
 var util = require('racer/lib/util');
 var derbyTemplates = require('derby-templates');
 var templates = derbyTemplates.templates;
 var expressions = derbyTemplates.expressions;
-var App = require('./App');
 var Controller = require('./Controller');
 
 exports.Component = Component;
 exports.ComponentFactory = ComponentFactory;
 exports.SingletonComponentFactory = SingletonComponentFactory;
 exports.createFactory = createFactory;
+exports.extendComponent = extendComponent;
 
 function Component(parent, context, id, scope) {
   this.parent = parent;
@@ -210,43 +209,6 @@ SingletonComponentFactory.prototype.init = function(context) {
 };
 // Don't call the create method for singleton components
 SingletonComponentFactory.prototype.create = function() {};
-
-App.prototype.component = function(viewName, constructor) {
-  if (typeof viewName === 'function') {
-    constructor = viewName;
-    viewName = null;
-  }
-
-  // Inherit from Component
-  extendComponent(constructor);
-
-  // Load template view from filename
-  if (constructor.prototype.view) {
-    var viewFilename = constructor.prototype.view;
-    viewName = constructor.prototype.name || path.basename(viewFilename, '.html');
-    this.loadViews(viewFilename, viewName);
-
-  } else if (!viewName) {
-    if (constructor.prototype.name) {
-      viewName = constructor.prototype.name;
-      var view = this.views.register(viewName);
-      view.template = templates.emptyTemplate;
-    } else {
-      throw new Error('No view name specified for component');
-    }
-  }
-
-  // Associate the appropriate view with the component type
-  var view = this.views.find(viewName);
-  if (!view) {
-    var message = this.views.findErrorMessage(viewName);
-    throw new Error(message);
-  }
-  view.componentFactory = createFactory(constructor);
-
-  // Make chainable
-  return this;
-};
 
 function extendComponent(constructor) {
   // Don't do anything if the constructor already extends Component
